Made it worse
Can't figure out the problem but I think it's in the array(s). The function that checks whether moves are valid or not think that all the moves are invalid for every location. I don't think it is looking at the part where the valid moves for each location are specified.